Addresses: Issue 14: MRAID object not available fast enough Issue 6: Display of IAB ad, Simple Resize (which was caused by Issue 14)
created a new function, loadHtml(), that is called once we know that mraid is available. This function loads the ad html into the iframe document and then cycles through any scripts (in adhtml), appending them to the body as needed. This functionality used to be in loadAd(), but would often fire before mraid was available.